feat(shared-resource): 资源详情页改为 GitHub Releases 式左右分栏
左侧 sidebar 列出全部版本（最新带标签、相对时间），右侧展示选中版本
的 tag 行、元数据与文件预览；选中同步 ?version=，刷新不丢、可直链。
版本详情主体抽成 SharedResourceVersionBody，与独立版本详情页共用，
活动/通知里的深链路入口保持不变。窄屏 sidebar 堆叠到详情上方。
